src/cron/requirements.txt
pika==1.3.2
loguru==0.7.2
pytz==2024.2
python-dotenv==1.0.1
